Add support for proper Gradle up-to-date checks
This adds a new UnMockTask class which defines the correct input
and output properties needed for Gradle to do proper up-to-date
checking on the task. It also allows Gradle to handle any file
downloads for dependent jars rather than trying to handle this
manually in ProcessRealAndroidJar.

In order to test that the caching logic is working correctly, add
a dependency on Gradle TestKit and java-gradle-plugin so that unit
tests can be written in the plugin project directly.

Fixes #46 and #47