Enable Gradle build cache across machines / checkouts
In the initial commit that added Gradle up-to-date check
and build cache support, we did not specify the @PathSensitive
annotation on File properties, which means that those task
inputs and outputs cannot be shared across multiple code
checkouts or across multiple machines.

Since the actual code of the task doesn't care what the names
or paths of the input files are, we can specify PathSensitivity.NONE
for all of them. This should give us the optimal number of cache
hits. Also refactored some of the plugin code to reduce duplicate file
path definitions.