Avoid dependency resolution at configuration time
Previously, the "unmock" configuration was resolved and
the robolectric dependency was downloaded at Gradle
configuration time, which would happen for ANY Gradle
invocation rather than only when the unmock task was
being run.

With this change, the default robolectric jar is
specified using Gradle's default dependency API so
we don't need to manually check the size of the configuration
and add it. We also pass the configuration FileCollection
to the task instead of passing the resolved file, which
lets resolution happen at task execution time.

Using the FileCollection#getSingleFile() API also gives
a much nicer error message if the user tries to add
multiple dependencies to the "unmock" configuration.
Previously we said to do a clean build or clear Gradle
caches, which wouldn't have actually fixed the problem.
Now the error (which comes from Gradle) explicitly says
that there should only be one dependency added to the
"unmock" configuration.