Only add task dependencies on compileUnitTest tasks
Previously the unmock tasks were added to every "compile"
task, which meant it would run even when compiling non-test
source code. Now it will only depend on the tasks that
compile JVM unit test code, not production code or androidTest
code.

We could potentially do even better and depend on the specific
tasks that compile JVM unit tests by depending on AGP and getting
a reference to the tasks explicitly instead of with a regex.
Depending on AGP, however, is kind of painful for consumers though,
since we would need unmock to update AGP in a timely manner when
new versions are released.